Refactor python layer.
