+++ pennylane_lightning/lightning_base.py
@@ -17,16 +17,15 @@ This module contains the base class for all PennyLane Lightning simulator device
 and interfaces with C++ for improved performance.
 """
 import numpy as np
+from pennylane.measurements import MeasurementProcess
 
 from ._version import __version__
 
 try:
     from .pennylane_lightning_ops import (
-        MeasurementsC64,
-        StateVectorC64,
-        MeasurementsC128,
-        StateVectorC128,
         backend_info,
+        MeasurementsC64,
+        MeasurementsC128,
     )
 
     CPP_BINARY_AVAILABLE = True
@@ -39,11 +38,24 @@ except ModuleNotFoundError:
 
 if CPP_BINARY_AVAILABLE:
     from typing import List
+    from itertools import product
 
     import pennylane as qml
     from pennylane import QubitDevice
     from pennylane.operation import Operation
     from pennylane.wires import Wires
+
+    from pennylane import (
+        BasisState,
+        QubitStateVector,
+    )
+
+    from .pennylane_lightning_ops.algorithms import (
+        create_ops_listC64,
+        create_ops_listC128,
+    )
+
+    from ._serialize import _serialize_observables, _serialize_ops
 
     class LightningBase(QubitDevice):
         """PennyLane Lightning Base device.
@@ -183,6 +195,202 @@ if CPP_BINARY_AVAILABLE:
             return super()._get_diagonalizing_gates(
                 qml.tape.QuantumScript(measurements=meas_filtered)
             )
+
+        def _preprocess_state_vector(self, state, device_wires):
+            """Initialize the internal state vector in a specified state.
+
+            Args:
+                state (array[complex]): normalized input state of length ``2**len(wires)``
+                    or broadcasted state of shape ``(batch_size, 2**len(wires))``
+                device_wires (Wires): wires that get initialized in the state
+
+            Returns:
+                array[complex]: normalized input state of length ``2**len(wires)``
+                    or broadcasted state of shape ``(batch_size, 2**len(wires))``
+                array[int]: indices for which the state is changed to input state vector elements
+            """
+
+            # translate to wire labels used by device
+            device_wires = self.map_wires(device_wires)
+
+            state = self._asarray(state, dtype=self.C_DTYPE)
+
+            if len(device_wires) == self.num_wires and Wires(sorted(device_wires)) == device_wires:
+                return None, state
+
+            # generate basis states on subset of qubits via the cartesian product
+            basis_states = np.array(list(product([0, 1], repeat=len(device_wires))))
+
+            # get basis states to alter on full set of qubits
+            unravelled_indices = np.zeros((2 ** len(device_wires), self.num_wires), dtype=int)
+            unravelled_indices[:, device_wires] = basis_states
+
+            # get indices for which the state is changed to input state vector elements
+            ravelled_indices = np.ravel_multi_index(unravelled_indices.T, [2] * self.num_wires)
+            return ravelled_indices, state
+
+        def _get_basis_state_index(self, state, wires):
+            """Returns the basis state index of a specified computational basis state.
+
+            Args:
+                state (array[int]): computational basis state of shape ``(wires,)``
+                    consisting of 0s and 1s
+                wires (Wires): wires that the provided computational state should be initialized on
+
+            Returns:
+                int: basis state index
+            """
+            # translate to wire labels used by device
+            device_wires = self.map_wires(wires)
+
+            # length of basis state parameter
+            n_basis_state = len(state)
+
+            if not set(state.tolist()).issubset({0, 1}):
+                raise ValueError("BasisState parameter must consist of 0 or 1 integers.")
+
+            if n_basis_state != len(device_wires):
+                raise ValueError("BasisState parameter and wires must be of equal length.")
+
+            # get computational basis state number
+            basis_states = 2 ** (self.num_wires - 1 - np.array(device_wires))
+            basis_states = qml.math.convert_like(basis_states, state)
+            return int(qml.math.dot(state, basis_states))
+
+        def _process_jacobian_tape(self, tape, starting_state, use_device_state):
+            # To support np.complex64 based on the type of self._state
+            if self.use_csingle:
+                create_ops_list = create_ops_listC64
+            else:
+                create_ops_list = create_ops_listC128
+
+            state_vector = self._init_process_jacobian_tape(tape, starting_state, use_device_state)
+
+            obs_serialized = _serialize_observables(
+                tape, self.wire_map, use_csingle=self.use_csingle
+            )
+            ops_serialized, use_sp = _serialize_ops(tape, self.wire_map)
+
+            ops_serialized = create_ops_list(*ops_serialized)
+
+            # We need to filter out indices in trainable_params which do not
+            # correspond to operators.
+            trainable_params = sorted(tape.trainable_params)
+            if len(trainable_params) == 0:
+                return None
+
+            tp_shift = []
+            record_tp_rows = []
+            all_params = 0
+
+            for op_idx, tp in enumerate(trainable_params):
+                # get op_idx-th operator among differentiable operators
+                op, _, _ = tape.get_operation(op_idx)
+                if isinstance(op, Operation) and not isinstance(op, (BasisState, QubitStateVector)):
+                    # We now just ignore non-op or state preps
+                    tp_shift.append(tp)
+                    record_tp_rows.append(all_params)
+                all_params += 1
+
+            if use_sp:
+                # When the first element of the tape is state preparation. Still, I am not sure
+                # whether there must be only one state preparation...
+                tp_shift = [i - 1 for i in tp_shift]
+
+            return {
+                "state_vector": state_vector,
+                "obs_serialized": obs_serialized,
+                "ops_serialized": ops_serialized,
+                "tp_shift": tp_shift,
+                "record_tp_rows": record_tp_rows,
+                "all_params": all_params,
+            }
+
+        @staticmethod
+        def _check_adjdiff_supported_measurements(measurements: List[MeasurementProcess]):
+            """Check whether given list of measurement is supported by adjoint_differentiation.
+
+            Args:
+                measurements (List[MeasurementProcess]): a list of measurement processes to check.
+
+            Returns:
+                Expectation or State: a common return type of measurements.
+            """
+            pass
+
+        @staticmethod
+        def _adjoint_jacobian_processing(jac):
+            """
+            Post-process the Jacobian matrix returned by ``adjoint_jacobian`` for
+            the new return type system.
+            """
+            jac = np.squeeze(jac)
+
+            if jac.ndim == 0:
+                return np.array(jac)
+
+            if jac.ndim == 1:
+                return tuple(np.array(j) for j in jac)
+
+            # must be 2-dimensional
+            return tuple(tuple(np.array(j_) for j_ in j) for j in jac)
+
+        def batch_vjp(
+            self, tapes, dys, reduction="append", starting_state=None, use_device_state=False
+        ):
+            """Generate the processing function required to compute the vector-Jacobian products
+            of a batch of tapes.
+
+            Args:
+                tapes (Sequence[.QuantumTape]): sequence of quantum tapes to differentiate
+                dys (Sequence[tensor_like]): Sequence of gradient-output vectors ``dy``. Must be the
+                    same length as ``tapes``. Each ``dy`` tensor should have shape
+                    matching the output shape of the corresponding tape.
+                reduction (str): Determines how the vector-Jacobian products are returned.
+                    If ``append``, then the output of the function will be of the form
+                    ``List[tensor_like]``, with each element corresponding to the VJP of each
+                    input tape. If ``extend``, then the output VJPs will be concatenated.
+                starting_state (tensor_like): post-forward pass state to start execution with. It should be
+                    complex-valued. Takes precedence over ``use_device_state``.
+                use_device_state (bool): use current device state to initialize. A forward pass of the same
+                    circuit should be the last thing the device has executed. If a ``starting_state`` is
+                    provided, that takes precedence.
+
+            Returns:
+                The processing function required to compute the vector-Jacobian products of a batch of tapes.
+            """
+            fns = []
+
+            # Loop through the tapes and dys vector
+            for tape, dy in zip(tapes, dys):
+                fn = self.vjp(
+                    tape.measurements,
+                    dy,
+                    starting_state=starting_state,
+                    use_device_state=use_device_state,
+                )
+                fns.append(fn)
+
+            def processing_fns(tapes):
+                vjps = []
+                for t, f in zip(tapes, fns):
+                    vjp = f(t)
+
+                    # make sure vjp is iterable if using extend reduction
+                    if (
+                        not isinstance(vjp, tuple)
+                        and getattr(reduction, "__name__", reduction) == "extend"
+                    ):
+                        vjp = (vjp,)
+
+                    if isinstance(reduction, str):
+                        getattr(vjps, reduction)(vjp)
+                    elif callable(reduction):
+                        reduction(vjps, vjp)
+
+                return vjps
+
+            return processing_fns
 
 else:  # binaries not available
     from pennylane.devices import DefaultQubit
