+++ pennylane_lightning/lightning_kokkos.py
@@ -19,11 +19,11 @@ interfaces with C++ for fast linear algebra calculations.
 import numpy as np
 from warnings import warn
 
-from .lightning_base import backend_info, CPP_BINARY_AVAILABLE, LightningBase
+from .lightning_base import backend_info, LightningBase
 
 if backend_info()["NAME"] == "lightning.kokkos":
     from typing import List
-    from itertools import islice, product
+    from itertools import islice
     from os import getenv
 
     from .pennylane_lightning_ops import (
@@ -299,6 +299,7 @@ if backend_info()["NAME"] == "lightning.kokkos":
 
         def _apply_state_vector(self, state, device_wires):
             """Initialize the internal state vector in a specified state.
+
             Args:
                 state (array[complex]): normalized input state of length ``2**len(wires)``
                     or broadcasted state of shape ``(batch_size, 2**len(wires))``
@@ -311,26 +312,16 @@ if backend_info()["NAME"] == "lightning.kokkos":
                 state.DeviceToHost(state_data.ravel(order="C"))
                 state = state_data
 
+            ravelled_indices, state = self._preprocess_state_vector(state, device_wires)
+
             # translate to wire labels used by device
             device_wires = self.map_wires(device_wires)
-
-            state = self._asarray(state, dtype=self.C_DTYPE)
             output_shape = [2] * self.num_wires
 
             if len(device_wires) == self.num_wires and Wires(sorted(device_wires)) == device_wires:
                 # Initialize the entire device state with the input state
                 self.syncH2D(self._reshape(state, output_shape))
                 return
-
-            # generate basis states on subset of qubits via the cartesian product
-            basis_states = np.array(list(product([0, 1], repeat=len(device_wires))))
-
-            # get basis states to alter on full set of qubits
-            unravelled_indices = np.zeros((2 ** len(device_wires), self.num_wires), dtype=int)
-            unravelled_indices[:, device_wires] = basis_states
-
-            # get indices for which the state is changed to input state vector elements
-            ravelled_indices = np.ravel_multi_index(unravelled_indices.T, [2] * self.num_wires)
 
             self._kokkos_state.setStateVector(ravelled_indices, state)  # this operation on device
 
@@ -344,23 +335,7 @@ if backend_info()["NAME"] == "lightning.kokkos":
 
             Note: This function does not support broadcasted inputs yet.
             """
-            # translate to wire labels used by device
-            device_wires = self.map_wires(wires)
-
-            # length of basis state parameter
-            n_basis_state = len(state)
-
-            if not set(state.tolist()).issubset({0, 1}):
-                raise ValueError("BasisState parameter must consist of 0 or 1 integers.")
-
-            if n_basis_state != len(device_wires):
-                raise ValueError("BasisState parameter and wires must be of equal length.")
-
-            # get computational basis state number
-            basis_states = 2 ** (self.num_wires - 1 - np.array(device_wires))
-            basis_states = qml.math.convert_like(basis_states, state)
-            num = int(qml.math.dot(state, basis_states))
-
+            num = self._get_basis_state_index(state, wires)
             self._create_basis_state(num)
 
         def apply_lightning(self, operations, apply_pre_rotated_state=False):
@@ -630,13 +605,7 @@ if backend_info()["NAME"] == "lightning.kokkos":
                         'the "adjoint" differentiation method'
                     )
 
-        def _process_jacobian_tape(self, tape, starting_state, use_device_state):
-            # To support np.complex64 based on the type of self._state
-            if self.use_csingle:
-                create_ops_list = create_ops_listC64
-            else:
-                create_ops_list = create_ops_listC128
-
+        def _init_process_jacobian_tape(self, tape, starting_state, use_device_state):
             # Initialization of state
             if starting_state is not None:
                 if starting_state.size != 2 ** len(self.wires):
@@ -648,47 +617,7 @@ if backend_info()["NAME"] == "lightning.kokkos":
             elif not use_device_state:
                 self.reset()
                 self.apply(tape.operations)
-
-            obs_serialized = _serialize_observables(
-                tape, self.wire_map, use_csingle=self.use_csingle
-            )
-            ops_serialized, use_sp = _serialize_ops(tape, self.wire_map)
-
-            ops_serialized = create_ops_list(*ops_serialized)
-
-            # We need to filter out indices in trainable_params which do not
-            # correspond to operators.
-            trainable_params = sorted(tape.trainable_params)
-            if len(trainable_params) == 0:
-                return None
-
-            tp_shift = []
-            record_tp_rows = []
-            all_params = 0
-
-            for op_idx, tp in enumerate(trainable_params):
-                # get op_idx-th operator among differentiable operators
-                op, _, _ = tape.get_operation(op_idx)
-                if isinstance(op, Operation) and not isinstance(op, (BasisState, QubitStateVector)):
-                    # We now just ignore non-op or state preps
-                    tp_shift.append(tp)
-                    record_tp_rows.append(all_params)
-                all_params += 1
-
-            if use_sp:
-                # When the first element of the tape is state preparation. Still, I am not sure
-                # whether there must be only one state preparation...
-                tp_shift = [i - 1 for i in tp_shift]
-
-            state_vector = self.state_vector
-            return {
-                "state_vector": state_vector,
-                "obs_serialized": obs_serialized,
-                "ops_serialized": ops_serialized,
-                "tp_shift": tp_shift,
-                "record_tp_rows": record_tp_rows,
-                "all_params": all_params,
-            }
+            return self.state_vector
 
         def adjoint_jacobian(self, tape, starting_state=None, use_device_state=False):
             if self.shots is not None:
@@ -750,23 +679,6 @@ if backend_info()["NAME"] == "lightning.kokkos":
             jac_r = np.zeros((jac.shape[0], processed_data["all_params"]))
             jac_r[:, processed_data["record_tp_rows"]] = jac
             return self._adjoint_jacobian_processing(jac_r) if qml.active_return() else jac_r
-
-        @staticmethod
-        def _adjoint_jacobian_processing(jac):
-            """
-            Post-process the Jacobian matrix returned by ``adjoint_jacobian`` for
-            the new return type system.
-            """
-            jac = np.squeeze(jac)
-
-            if jac.ndim == 0:
-                return np.array(jac)
-
-            if jac.ndim == 1:
-                return tuple(np.array(j) for j in jac)
-
-            # must be 2-dimensional
-            return tuple(tuple(np.array(j_) for j_ in j) for j in jac)
 
         def vjp(self, measurements, dy, starting_state=None, use_device_state=False):
             """Generate the processing function required to compute the vector-Jacobian products of a tape.
@@ -838,63 +750,6 @@ if backend_info()["NAME"] == "lightning.kokkos":
                     return self.adjoint_jacobian(new_tape, starting_state, use_device_state)
 
                 return processing_fn
-
-        def batch_vjp(
-            self, tapes, dys, reduction="append", starting_state=None, use_device_state=False
-        ):
-            """Generate the processing function required to compute the vector-Jacobian products
-            of a batch of tapes.
-
-            Args:
-                tapes (Sequence[.QuantumTape]): sequence of quantum tapes to differentiate
-                dys (Sequence[tensor_like]): Sequence of gradient-output vectors ``dy``. Must be the
-                    same length as ``tapes``. Each ``dy`` tensor should have shape
-                    matching the output shape of the corresponding tape.
-                reduction (str): Determines how the vector-Jacobian products are returned.
-                    If ``append``, then the output of the function will be of the form
-                    ``List[tensor_like]``, with each element corresponding to the VJP of each
-                    input tape. If ``extend``, then the output VJPs will be concatenated.
-                starting_state (tensor_like): post-forward pass state to start execution with. It should be
-                    complex-valued. Takes precedence over ``use_device_state``.
-                use_device_state (bool): use current device state to initialize. A forward pass of the same
-                    circuit should be the last thing the device has executed. If a ``starting_state`` is
-                    provided, that takes precedence.
-
-            Returns:
-                The processing function required to compute the vector-Jacobian products of a batch of tapes.
-            """
-            fns = []
-
-            # Loop through the tapes and dys vector
-            for tape, dy in zip(tapes, dys):
-                fn = self.vjp(
-                    tape.measurements,
-                    dy,
-                    starting_state=starting_state,
-                    use_device_state=use_device_state,
-                )
-                fns.append(fn)
-
-            def processing_fns(tapes):
-                vjps = []
-                for t, f in zip(tapes, fns):
-                    vjp = f(t)
-
-                    # make sure vjp is iterable if using extend reduction
-                    if (
-                        not isinstance(vjp, tuple)
-                        and getattr(reduction, "__name__", reduction) == "extend"
-                    ):
-                        vjp = (vjp,)
-
-                    if isinstance(reduction, str):
-                        getattr(vjps, reduction)(vjp)
-                    elif callable(reduction):
-                        reduction(vjps, vjp)
-
-                return vjps
-
-            return processing_fns
 
 else:
